Update featuring optimisations
The next three months, I had to create a website for my project with a well-based front-end and back-end, test the product in my room almost daily, fix the bugs and take my exams.
This is the piece I updated from the last source code with further optimisation and bug fixes.

In the event of winning the competition in the county, I will have to do more upgrades to it, but we'll see what the future gives in.
- 19th of April 2017