Faster myHaum and gas calibration
Had to calibrate the gas sensor because it was always feeling gas when the parameters weren't set right. Meanwhile, we found a hack in which the ESP WiFi module connects automatically to the last WiFi it was set onto. The results amazed everyone, it connected to WiFi almost instantaneously, with no waiting screen.